\name{GetWCSTemplate}
\alias{GetWCSTemplate}
\title{A function to return the OGC Web Coverage Request parameter template as a data frame}
\usage{
GetWCSTemplate()
}
\value{
a data frame with three columns: param, type, value (NAs)
containing the OGC WCS parameters
}
\description{
A function to return the OGC Web Coverage Request parameter
template as a data frame
}
\examples{
\dontrun{
GetWCSTemplate()
}
\keyword{utilities}

